requirements.txt
discord.py
opencv-python